This might work
Hopefully this will work. I think the problem before was that github was looking into the Julianns's Website file that had nothing on it and now it is hopefully maybe looking at the correct file, because I just moved the one with all the code in it into the one that I think github is referencing.